DCIO-pipeline/requirements.txt
pdfplumber==0.11.4
camelot-py==0.11.0
ghostscript==0.7
pdf2image==1.16.3
pillow>=11.0.0
numpy==1.26.4
pandas==2.2.1
sqlalchemy==2.0.28
pydantic>=2.7.0
python-dotenv==1.0.1
openai==1.50.2
rapidfuzz==3.6.1
pyyaml==6.0.2
